511578 Isuzu Plant 5
LsdTc: Added VL36 versions of the LSD test, as for the new vehicle the standard axle does not spin the opposite wheel during the test. This was causing the test to fail when it was actually working as intended according to the manufacturer.

AllisonTransmissionTC: 
Fixed a bug that caused the shift lever test to always report as "skipped" on a retest. 

IsuzuEmissionsTc:
Enabled the module lock test to be skipped with a parameter.


git-svn-id: http://10.105.1.46/svn/CCRT/Ccrt/branches/Common3700@30396 80089e14-fa6c-c64e-badd-1f0333e23b03